Add element interface library and core documentation
This is a hefty update, but it is aimed at the finish line on Kava's primary goals. Like Bumblebee and Ironhide, Kava is meant to be a POM wrapper on the consumed test framework, facading system interactions into a highly-readable, concrete interface. In comparison, XCTest is not very readable when dealing with complex interactions, and the framework is as dynamic as you can get (every XCUIElement has methods for every possible interaction in the system). At this point, we will only be adding common interactions and Element subclasses, but this should handle 95% of any iOS application UI.

Reviewed by Brett and Gilbert

https://github.com/mindbody/Kava/pull/1